package optimizations
