+++ Alchemy/neural_networks/convolutional_neural_network_tf.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 from tensorflow import keras
 from tensorflow.keras import layers
-from MaciNet.utils import train_test_split, to_categorical, Plot
+from Alchemy.utils import train_test_split, to_categorical, Plot
 
 
 def build_convolutional_neural_network_tf(datapath:str, plot_results:bool=True):
