+++ CoFEA/models_rl/sarsa_fea.py
@@ -5,11 +5,11 @@ import numpy as np
 from .params import *
 from .agent import Agent
 from CoFEA.environments import environment
-from MaciNet.supervised_learning import ParticleSwarmOptimizedNN
-from MaciNet.deep_learning import NeuralNetwork
-from MaciNet.deep_learning.layers import Activation, Dense
-from MaciNet.deep_learning.loss_functions import CrossEntropy
-from MaciNet.deep_learning.optimizers import Adam
+from Alchemy.supervised_learning import ParticleSwarmOptimizedNN
+from Alchemy.deep_learning import NeuralNetwork
+from Alchemy.deep_learning.layers import Activation, Dense
+from Alchemy.deep_learning.loss_functions import CrossEntropy
+from Alchemy.deep_learning.optimizers import Adam
 import gym
